.github/workflows/scripts/requirements.txt
pyyaml>=6.0
GitPython>=3.1.0
loguru>=0.7.0
